requirements.txt
-e .

click==8.1.3
hydra-core==1.3.1
kaggle==1.5.12
matplotlib==3.5.1
numpy==1.21.2
omegaconf==2.3.0
pandas==1.5.2
Pillow==9.4.0
python-dotenv==0.21.0
pytorch_lightning==1.8.6
scikit_learn==1.2.0
setuptools==65.6.3
torch==1.11.0
torchvision==0.12.0
